+++ src/test/java/dinkplugin/notifiers/DeathNotifierTest.java
@@ -10,9 +10,14 @@ import net.runelite.api.InventoryID;
 import net.runelite.api.Item;
 import net.runelite.api.ItemContainer;
 import net.runelite.api.ItemID;
+import net.runelite.api.NPC;
+import net.runelite.api.NPCComposition;
+import net.runelite.api.NpcID;
+import net.runelite.api.ParamID;
 import net.runelite.api.Player;
 import net.runelite.api.Prayer;
 import net.runelite.api.Varbits;
+import net.runelite.api.coords.LocalPoint;
 import net.runelite.api.coords.WorldPoint;
 import net.runelite.api.events.ActorDeath;
 import net.runelite.api.events.InteractingChanged;
@@ -26,6 +31,7 @@ import java.util.List;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyBoolean;
+import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
@@ -59,8 +65,10 @@ class DeathNotifierTest extends MockedNotifierTest {
         // init client mocks
         when(client.getVarbitValue(Varbits.IN_WILDERNESS)).thenReturn(1);
         when(client.getPlayers()).thenReturn(Collections.emptyList());
+        when(client.getCachedNPCs()).thenReturn(new NPC[0]);
         WorldPoint location = new WorldPoint(0, 0, 0);
         when(localPlayer.getWorldLocation()).thenReturn(location);
+        when(localPlayer.getLocalLocation()).thenReturn(new LocalPoint(0, 0));
 
         // init item mocks
         mockItem(ItemID.RUBY, RUBY_PRICE, "Ruby");
@@ -68,6 +76,9 @@ class DeathNotifierTest extends MockedNotifierTest {
         mockItem(ItemID.OPAL, OPAL_PRICE, "Opal");
         mockItem(ItemID.COAL, COAL_PRICE, "Coal");
         mockItem(ItemID.TUNA, TUNA_PRICE, "Tuna");
+
+        // init npc mocks
+        when(npcManager.getHealth(anyInt())).thenReturn(50);
     }
 
     @Test
@@ -81,7 +92,7 @@ class DeathNotifierTest extends MockedNotifierTest {
             false,
             NotificationBody.builder()
                 .text(String.format("%s has died, losing %d gp", PLAYER_NAME, 0))
-                .extra(new DeathNotificationData(0L, false, null, Collections.emptyList(), Collections.emptyList()))
+                .extra(new DeathNotificationData(0L, false, null, null, null, Collections.emptyList(), Collections.emptyList()))
                 .type(NotificationType.DEATH)
                 .build()
         );
@@ -126,7 +137,7 @@ class DeathNotifierTest extends MockedNotifierTest {
             false,
             NotificationBody.builder()
                 .text(String.format("%s has died, losing %d gp", PLAYER_NAME, TUNA_PRICE))
-                .extra(new DeathNotificationData((long) TUNA_PRICE, false, null, kept, lost))
+                .extra(new DeathNotificationData(TUNA_PRICE, false, null, null, null, kept, lost))
                 .type(NotificationType.DEATH)
                 .embeds(embeds)
                 .build()
@@ -151,7 +162,7 @@ class DeathNotifierTest extends MockedNotifierTest {
             false,
             NotificationBody.builder()
                 .text(String.format("%s has been PKed by %s for %d gp", PLAYER_NAME, pker, 0))
-                .extra(new DeathNotificationData(0L, true, pker, Collections.emptyList(), Collections.emptyList()))
+                .extra(new DeathNotificationData(0L, true, pker, pker, null, Collections.emptyList(), Collections.emptyList()))
                 .type(NotificationType.DEATH)
                 .build()
         );
@@ -178,7 +189,7 @@ class DeathNotifierTest extends MockedNotifierTest {
             false,
             NotificationBody.builder()
                 .text(String.format("%s has been PKed by %s for %d gp", PLAYER_NAME, pker, 0))
-                .extra(new DeathNotificationData(0L, true, pker, Collections.emptyList(), Collections.emptyList()))
+                .extra(new DeathNotificationData(0L, true, pker, pker, null, Collections.emptyList(), Collections.emptyList()))
                 .type(NotificationType.DEATH)
                 .build()
         );
@@ -202,7 +213,49 @@ class DeathNotifierTest extends MockedNotifierTest {
             false,
             NotificationBody.builder()
                 .text(String.format("%s has died, losing %d gp", PLAYER_NAME, 0))
-                .extra(new DeathNotificationData(0L, false, null, Collections.emptyList(), Collections.emptyList()))
+                .extra(new DeathNotificationData(0L, false, null, null, null, Collections.emptyList(), Collections.emptyList()))
+                .type(NotificationType.DEATH)
+                .build()
+        );
+    }
+
+    @Test
+    void testNotifyNpc() {
+        // init mocks
+        String name = "Guard";
+        NPC other = mock(NPC.class);
+        when(other.getName()).thenReturn(name);
+        when(other.getId()).thenReturn(NpcID.GUARD);
+        when(other.isDead()).thenReturn(false);
+        when(other.getCombatLevel()).thenReturn(21);
+        when(other.getInteracting()).thenReturn(localPlayer);
+        when(other.getLocalLocation()).thenReturn(new LocalPoint(1, 1));
+
+        NPCComposition comp = mock(NPCComposition.class);
+        when(other.getTransformedComposition()).thenReturn(comp);
+        when(comp.isInteractible()).thenReturn(true);
+        when(comp.isFollower()).thenReturn(false);
+        when(comp.getSize()).thenReturn(1);
+        when(comp.isMinimapVisible()).thenReturn(true);
+        when(comp.getId()).thenReturn(NpcID.GUARD);
+        when(comp.getName()).thenReturn(name);
+        when(comp.getStringValue(ParamID.NPC_HP_NAME)).thenReturn(name);
+        when(comp.getCombatLevel()).thenReturn(21);
+        when(comp.getActions()).thenReturn(new String[] { "Pickpocket", "Attack", "Examine" });
+
+        when(npcManager.getHealth(NpcID.GUARD)).thenReturn(22);
+        when(client.getCachedNPCs()).thenReturn(new NPC[] { other });
+
+        // fire event
+        plugin.onActorDeath(new ActorDeath(localPlayer));
+
+        // verify notification
+        verify(messageHandler).createMessage(
+            PRIMARY_WEBHOOK_URL,
+            false,
+            NotificationBody.builder()
+                .text(String.format("%s has died, losing %d gp", PLAYER_NAME, 0))
+                .extra(new DeathNotificationData(0L, false, null, name, NpcID.GUARD, Collections.emptyList(), Collections.emptyList()))
                 .type(NotificationType.DEATH)
                 .build()
         );